move ui commande to ui folder and add positional
